Update rssim.cpp
Based on the Equations document (as well as textbooks I've checked), I believe signs inside the square roots of the Doppler equation should be swapped? I.e. Addition in the numerator and subtraction in the numerator? This form is also seen in the SolveREDirect function, so I assume it's a mistake in SolveRE?

When testing the current version using a monostatic setup, the Doppler shift is computed as negative for an approaching target. This cannot be correct as it implies a negative target velocity towards the radar. So either the signs need to be swapped in the equation above, or (Fc - Fr) in "rsresponse.cpp" should be changed to (Fr - Fc) such that the Doppler is correct.